+++ Modules/bulletjournal/Controller/BjnotesController.php
@@ -5,6 +5,9 @@ require_once 'Framework/Form.php';
 require_once 'Modules/core/Controller/CoresecureController.php';
 require_once 'Modules/bulletjournal/Model/BulletjournalTranslator.php';
 require_once 'Modules/bulletjournal/Model/BjNote.php';
+require_once 'Modules/bulletjournal/Model/BjEvent.php';
+require_once 'Modules/bulletjournal/Model/BjTask.php';
+
 require_once 'Modules/core/Controller/CorespaceController.php';
 
 /**
@@ -112,4 +115,16 @@ class BjnotesController extends CoresecureController {
         $form->setValidationButton(CoreTranslator::Save($lang), "bjeditevent/".$id_space);
         return $form->getHtml($lang);
     }
+
+    public function deletenoteAction($id_space, $id) {
+        $this->checkAuthorizationMenuSpace("bulletjournal", $id_space, $_SESSION["id_user"]);
+        $modelNote = new BjNote();
+        $modelNote->delete($id_space, $id);
+        $modelEvent = new BjEvent();
+        $modelEvent->delete($id_space, $id);
+        $modelTask = new BjTask();
+        $modelTask->delete($id_space, $id);
+        return null;
+    }
+
 }